backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Firebase
firebase-admin==6.5.0

# Google Auth (for Firebase JWT verification)
google-auth==2.35.0

# Google Gemini SDK
google-generativeai==0.8.3

# OpenAI SDK (for DALL-E)
openai==1.54.4

# Async HTTP client (replaces fetch)
httpx==0.27.2

# Data validation (built into FastAPI)
pydantic==2.9.2
pydantic-settings==2.6.1

# File parsing
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5

# File upload support
python-multipart==0.0.12

# Environment variables
python-dotenv==1.0.1

# CORS
starlette==0.41.2

# ── Phase 1: Enterprise Foundation ──────────────────────────────────
# Database (Postgres + SQLAlchemy async)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Cache + Message Queue
redis[hiredis]==5.2.1
celery[redis]==5.4.0

# Access Control
casbin==1.36.2

# Notifications
sendgrid==6.11.0
slack-sdk==3.33.4
jinja2==3.1.4

# Observability
structlog==24.4.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
